+++ History.md
@@ -1,3 +1,25 @@
+WORKING
+===================
+  * TODO: Add animation class to chLib js. 
+
+1.0.4 / 2017-03-01
+===================
+  * deps: mouseManager
+    - Fix: Safari and Chrome mobile emulation fires mouse touch and mouseDown, generating twice clic.
+           Add flag to prevent this happend.
+
+  * deps: canvasControl
+    - Fix: Same logic from mouseManager twice clic.
+
+1.0.3 / 2016-18-10
+===================
+  * deps: ControlCanvas@3.6
+    - Fix: Mouse not working when canvas is scaled
+    - Refactor: no more canvas, it was replace for canvasEx, this solved mouse scale problems
+
+  * deps: ChLib@1.1
+    - Fix: ChCanvas class now has m_scaleX, m_scaleY setted properly.
+
 1.0.2 / 2016-09-05
 ===================
   * deps: ControlCanvas@3.5